src/requirements.txt
backcall==0.1.0
click==6.7
decorator==4.3.0
Django==2.0.6
django-cors-headers==2.1.0
django-progressive-web-app==0.1.1
django-redis==4.8.0
djangorestframework==3.8.2
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
mysqlclient==1.3.10
parso==0.3.0
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-dotenv==0.6.4
pytz==2017.2
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
